pyproject.toml
[project]
name = "coastlines-stats"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dep-tools",
    "exactextract>=0.2.2",
    "gdal==3.8.4",
    "geopandas>=1.1.1",
    "odc-geo>=0.4.10",
    "odc-stac>=0.4.0",
    "pyarrow>=21.0.0",
    "pyogrio>=0.11.1",
    "pystac-client>=0.9.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
]

[tool.uv.sources]
dep-tools = { git = "https://github.com/digitalearthpacific/dep-tools.git" }
